Character Reorganisation and CombatManagers.
Removed SheetCharacter and SimpleCharacter, all characters contain a details field which can contain a CharacterSheet or SimpleCharacterDetails.

Characters can now be ServerCharacter or ClientCharacter to split up logic similarly to Server/ClientPlayerEntitiy.

Gave the CombatManager some functionality so that characters become part of combat when they attack each other, and characters can be removed from combat with the /dnd flee <character> command.

+++ src/main/kotlin/com/pineypiney/mtt/serialisation/MTTCodecs.kt
@@ -8,12 +8,12 @@ import kotlin.math.min
 object MTTCodecs {
 
 
-	fun createBoolByte(vararg values: Boolean): Byte {
+	fun createBoolByte(vararg values: Boolean): Int {
 		var b = 0
 		for (i in 0..<min(values.size, 8)) {
 			b = b or if (values[i]) 1 shl i else 0
 		}
-		return b.toByte()
+		return b
 	}
 
 	val UUID_CODEC: Codec<UUID> = RecordCodecBuilder.create { builder ->
